replace int(round(...)) with round(...)
We don't need to cast to int when using the round function from py23,
as this is a backport of python3's built-in round and thus it returns
an int when called with a single argument.

+++ Lib/fontTools/misc/fixedTools.py
@@ -47,9 +47,9 @@ def fixedToFloat(value, precisionBits):
 
 def floatToFixed(value, precisionBits):
 	"""Converts a float to a fixed-point number given the number of
-	precisionBits.  Ie. int(round(value * (1<<precisionBits))).
+	precisionBits.  Ie. round(value * (1<<precisionBits)).
 	"""
-	return int(round(value * (1<<precisionBits)))
+	return round(value * (1<<precisionBits))
 
 def floatToFixedToFloat(value, precisionBits):
 	"""Converts a float to a fixed-point number given the number of
